Rename confusing selector.ts file
