+++ src/contentScript/nativeEditor/elementPicker.ts
