+++ src/contentScript/nativeEditor/insertPanel.tsx
@@ -18,7 +18,7 @@
 // https://github.com/facebook/react/blob/7559722a865e89992f75ff38c1015a865660c3cd/packages/react-devtools-shared/src/backend/views/Highlighter/index.js
 
 import { uuidv4 } from "@/types/helpers";
-import { userSelectElement } from "./selector";
+import { userSelectElement } from "./elementPicker";
 import * as pageScript from "@/pageScript/protocol";
 import { findContainer } from "./selectorInference";
 import { html as beautifyHTML } from "js-beautify";
